+++ spec/cookbooks/test/metadata.rb
@@ -4,3 +4,4 @@ name		'test'
 version		'0.0.1'
 
 depends		'zap', '>= 0.7.0'
+depends		'cron'